ゲーム編集画面を実装
view の共通部分を views/game/Form にまとめた
state の共通部分を state/game/gameForm にまとめた
List の UI を少し変更した

+++ src/views/game/List/CardElem.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react'
-import { makeStyles, colors, Grid, Card, CardContent, CardActions, IconButton } from '@material-ui/core'
-import { PlayCircleFilled, Delete } from '@material-ui/icons'
+import { makeStyles, colors, Grid, Card, CardContent, CardActions, IconButton, Menu, MenuItem } from '@material-ui/core'
+import { PlayCircleFilled, Delete, MoreHoriz, Edit } from '@material-ui/icons'
 import { useHistory } from 'react-router-dom'
 import { useAppDispatch, thunkActions } from 'state'
 
@@ -17,13 +17,16 @@ const useStales = makeStyles({
   grow: {
     flexGrow: 1,
   },
+  space: {
+    marginLeft: 8
+  }
 })
 
 type Props = {
   index: string
 }
 
-export const CardElem: React.FC<Props> = ({index, children}) => {
+export const CardElem: React.FC<Props> = ({ index, children }) => {
   const history = useHistory()
   const link = (index: string) => () => {
     history.push('/games/' + index)
@@ -32,10 +35,29 @@ export const CardElem: React.FC<Props> = ({index, children}) => {
   const classes = useStales()
   const dispatch = useAppDispatch()
 
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
+  const open = Boolean(anchorEl)
+  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget)
+  }
+  const handleClose = () => {
+    setAnchorEl(null)
+  }
+
   const deleteGame = useCallback(
     () => {
-      dispatch(thunkActions.gameList.deleteGame(index))
+      if (window.confirm('削除しますか')) {
+        handleClose()
+        dispatch(thunkActions.gameList.deleteGame(index))
+      }
     }, [dispatch, index]
+  )
+
+  const editGame = useCallback(
+    () => {
+      handleClose()
+      history.push('/games/edit/' + index)
+    }, [history, index]
   )
 
   return (
@@ -43,12 +65,32 @@ export const CardElem: React.FC<Props> = ({index, children}) => {
       <Card className={classes.card}>
         <CardActions>
           <IconButton color='primary' onClick={link(index)}>
-            <PlayCircleFilled/>
+            <PlayCircleFilled />
           </IconButton>
-          <div className={classes.grow}/>
-          <IconButton color='primary' onClick={deleteGame}>
-            <Delete/>
+          <div className={classes.grow} />
+          <IconButton
+            color='primary'
+            aria-label="more"
+            aria-controls="card-menu"
+            aria-haspopup="true"
+            onClick={handleClick}
+          >
+            <MoreHoriz />
           </IconButton>
+          <Menu
+            id="long-menu"
+            anchorEl={anchorEl}
+            keepMounted
+            open={open}
+            onClose={handleClose}
+          >
+            <MenuItem onClick={editGame}>
+              <Edit /> <span className={classes.space}/> 編集
+            </MenuItem>
+            <MenuItem onClick={deleteGame}>
+              <Delete /> <span className={classes.space}/> 削除
+            </MenuItem>
+          </Menu>
         </CardActions>
         <CardContent className={classes.cardContent}>
           {children}
